Working on do the train 0.0.2(e)
cont updating ui and little fix in biz logic
