+++ lib/features/train/do_the_train/data/repo/do_the_train_data_repo_impl.dart
@@ -9,4 +9,5 @@ abstract class DoTheTrainDataRepoImpl {
   Future<void> nextExercise({required TempTrainModel train});
   Future<List<ExerciseModel>> getProgressTempTrain(
       {required List<String> exercisesList});
+  Future<void> exitFromTrainWhenTrainIsOver();
 }
