+++ lib/features/train/do_the_train/domain/repo/do_the_train_domain_repo_impl.dart
@@ -5,4 +5,5 @@ abstract class DoTheTrainDomainRepoImpl {
   Future<bool> completeTrainAndExit({required TempTrainModel train});
   Future<void> nextExercise({required TempTrainModel train});
   Future<bool> exitFromTrainWithoutSaving();
+  Future<void> exitFromTrainingWhenAppCrash();
 }
