+++ lib/features/train/do_the_train/presentation/view_result/view_good_result.dart
@@ -1,0 +1,1 @@
+do this
