+++ lib/features/train/do_the_train/presentation/view_temp_progress_train/components/back_to_train_in_view_progress.dart
@@ -1,0 +1,41 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class BackToTrainButtonInViewProgress extends StatelessWidget {
+  const BackToTrainButtonInViewProgress({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 15),
+      child: Container(
+        decoration: BoxDecoration(
+            borderRadius: const BorderRadius.all(Radius.circular(99)),
+            gradient: LinearGradient(colors: [
+              Theme.of(context).colorScheme.primaryFixedDim,
+              Theme.of(context).colorScheme.secondaryFixedDim,
+            ], transform: const GradientRotation(pi / 2))),
+        child: ElevatedButton(
+            onPressed: () {
+              context.pop();
+            },
+            style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
+                fixedSize: WidgetStatePropertyAll(
+                    Size(MediaQuery.of(context).size.width, 60)),
+                backgroundColor:
+                    const WidgetStatePropertyAll(Colors.transparent)),
+            child: Text('Назад',
+                style: GoogleFonts.inter(
+                    fontSize: 18,
+                    fontWeight: FontWeight.w500,
+                    color: Theme.of(context).colorScheme.onSecondary))),
+      ),
+    );
+  }
+}
