+++ lib/features/train/do_the_train/presentation/view_temp_progress_train/components/count_of_reps_in_view_progress.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class CoutOfRepsInfoInViewProgress extends StatelessWidget {
+  const CoutOfRepsInfoInViewProgress(
+      {super.key, required this.coutOfRepInThatExercise});
+
+  final String? coutOfRepInThatExercise;
+
+  @override
+  Widget build(BuildContext context) {
+    return Text('Количество подходов: $coutOfRepInThatExercise',
+        style: GoogleFonts.inter(
+            fontSize: 15,
+            fontWeight: FontWeight.w500,
+            color: Theme.of(context).colorScheme.onSecondary));
+  }
+}
