+++ lib/navigation/router.dart
@@ -33,9 +33,9 @@ import 'package:fitflow/features/home/future_profile/presentation/indicators/wei
 import 'package:fitflow/features/loading/presentation/loading_main_widget.dart';
 import 'package:fitflow/features/train/create_training_plan/presentation/view_done_plan/edit_day_in_plan/edit_day_in_plan_main_widget.dart';
 import 'package:fitflow/features/train/create_training_plan/presentation/view_done_plan/view_done_plan_main_widget.dart';
-import 'package:fitflow/features/train/do_the_train/presentation/do_the_train_main_widget.dart';
-import 'package:fitflow/features/train/do_the_train/presentation/exit_the_train.dart';
-import 'package:fitflow/features/train/do_the_train/presentation/view_temp_progress.dart';
+import 'package:fitflow/features/train/do_the_train/presentation/train/do_the_train_main_widget.dart';
+import 'package:fitflow/features/train/do_the_train/presentation/exit_train_before_end/exit_the_train.dart';
+import 'package:fitflow/features/train/do_the_train/presentation/view_temp_progress_train/view_temp_progress.dart';
 import 'package:fitflow/navigation/home_navigation_bar/navbar.dart';
 import 'package:fitflow/navigation/paths.dart';
 import 'package:flutter/material.dart';
